+++ reportportal_client/helpers/__init__.py
@@ -4,6 +4,10 @@ from reportportal_client.helpers import markdown_helpers
 from reportportal_client.helpers.common_helpers import (
     ATTRIBUTE_LENGTH_LIMIT,
     CONTENT_TYPE_TO_EXTENSIONS,
+    ITEM_DESCRIPTION_LENGTH_LIMIT,
+    ITEM_NAME_LENGTH_LIMIT,
+    LAUNCH_DESCRIPTION_LENGTH_LIMIT,
+    LAUNCH_NAME_LENGTH_LIMIT,
     TRUNCATE_REPLACEMENT,
     TYPICAL_FILE_PART_HEADER,
     TYPICAL_MULTIPART_FOOTER_LENGTH,
@@ -36,6 +40,10 @@ from reportportal_client.helpers.common_helpers import (
 __all__ = [
     "markdown_helpers",
     "ATTRIBUTE_LENGTH_LIMIT",
+    "LAUNCH_DESCRIPTION_LENGTH_LIMIT",
+    "ITEM_DESCRIPTION_LENGTH_LIMIT",
+    "LAUNCH_NAME_LENGTH_LIMIT",
+    "ITEM_NAME_LENGTH_LIMIT",
     "TRUNCATE_REPLACEMENT",
     "CONTENT_TYPE_TO_EXTENSIONS",
     "TYPICAL_MULTIPART_FOOTER_LENGTH",